refactor(zai): give the top-level reasoning_effort field its own module
The file-size gate rejected the previous commit: `zai.rs` reached 1636 lines
against a grandfathered ceiling of 1572, and `zai/tests.rs` 1932 against 1843.
The gate offers `make file-size-update` only when the growth is irreducible and
otherwise asks that new code go in its own module. Two pure functions and their
tests are reducible, so this takes the second branch and leaves the baseline
alone.

The split follows a real seam rather than shedding lines wherever they fell.
`zai/effort.rs` holds the identities that carry reasoning depth in OpenAI's
top-level `reasoning_effort` string — xAI, for which it is the only reasoning
control, and Z.ai, for which it is depth alongside GLM's `thinking` on/off
object. OpenRouter stays behind deliberately: it expresses both on/off and depth
in a structured `reasoning` object, which belongs with the request body.

The test split mirrors `openrouter_effort_tests.rs`, whose own header records
that it was carved off this same file for this same gate.

Consequences worth naming:

* the five helpers need `pub(super)`, since a child module's items are private
  to its parent;
* only `map_zai_effort` is re-exported under `#[cfg(test)]`. Its xAI sibling is
  NOT, and the asymmetry is real rather than an oversight: the xAI tests assert
  on `xai_supports_reasoning_effort`, the model-gating predicate, and never on
  the mapping table, so re-exporting it would fail clippy's `-D warnings` as an
  unused import. `cargo test` does not enforce that lint, so a green test run is
  not evidence this compiles under the gate — run clippy;
* `provider_parity`'s `adapter_sources()` had to learn the new file. That guard
  reads source TEXT rather than the module tree, so a witness that moves into a
  split-out module vanishes from it and fails — which its own doc comment
  predicts, noting the list has to follow the splits the ratchet keeps forcing.

zai.rs 1636 -> 1533, zai/tests.rs 1932 -> 1814. No behaviour change:
stella-model 353 tests pass including the parity witness guard, stella-cli 1281,
and `clippy --all-targets -- -D warnings` is clean on both.
