+++ crates/stella-model/src/provider_parity.rs
@@ -310,7 +310,7 @@ mod tests {
     /// is a false alarm rather than the rotted proof it exists to catch. The
     /// parent `tests.rs` is over the file-size ratchet, so those splits keep
     /// happening — the list has to follow them.
-    fn adapter_sources() -> [&'static str; 12] {
+    fn adapter_sources() -> [&'static str; 13] {
         [
             include_str!("anthropic/tests.rs"),
             include_str!("anthropic/tests/thinking.rs"),
@@ -323,6 +323,7 @@ mod tests {
             include_str!("zai/tests/openrouter_effort_tests.rs"),
             include_str!("zai/tests/openrouter_stream_tests.rs"),
             include_str!("zai/tests/stream_frame_tests.rs"),
+            include_str!("zai/tests/zai_effort_tests.rs"),
             include_str!("zai/vision_tests.rs"),
         ]
     }
